Did my first task (1a)
1a. Spot and holes
The players can play the game with the basic configuration of the grid
in which Spot can move and fall into holes, losing a life each time this
happens until there are no lives left. This requires:
positioning Spot and 12 holes at random on the grid, making sure they
are not placed on the same cells
allowing the player to control Spot's movements on the grid
ensuring that Spot stays within the grid's boundaries, and remains in
same place if the player tries to move him onto that wall
detecting when Spot has fallen into a hole and reacting as specified
(see rules)
indicating at any time the number of remaining lives
terminating the game, with an appropriate message, (a) when the player
wants to quit (e.g., "YOU QUIT!") or (b) when Spot has lost all his
allocated lives (e.g., "YOU LOST!").